test: introduce LiveApp class to interact mainly with dummy app on mobile
got initial message from webview
